Refactor out account controller methods
