+++ app/Http/Controllers/Adm/Mship/Account/Roles.php
@@ -1,0 +1,63 @@
+<?php
+
+namespace App\Http\Controllers\Adm\Mship\Account;
+
+use App\Http\Controllers\Adm\AdmController;
+use App\Models\Mship\Account as AccountData;
+use Auth;
+use Input;
+use Redirect;
+use Session;
+use Spatie\Permission\Models\Role as RoleData;
+use URL;
+
+class Roles extends AdmController
+{
+    public function getRoleDetach(AccountData $mshipAccount, RoleData $role)
+    {
+        if (!$mshipAccount) {
+            return Redirect::route('adm.mship.account.index');
+        }
+
+        if (!$role) {
+            return Redirect::route('adm.mship.account.details', [$mshipAccount->id, 'roles'])
+                ->withError('The selected role does not exist.');
+        }
+
+        if (!$mshipAccount->roles->contains($role->id)) {
+            return Redirect::route('adm.mship.account.details', [$mshipAccount->id, 'roles'])
+                ->withError('This role is not attached to this user.');
+        }
+
+        // Let's remove!
+        $mshipAccount->roles()
+            ->detach($role);
+
+        return Redirect::route('adm.mship.account.details', [$mshipAccount->id, 'roles'])
+            ->withSuccess($role->name . ' role detached successfully. This user lost ' . count($role->permissions) . ' permissions.');
+    }
+
+    public function postRoleAttach(AccountData $mshipAccount)
+    {
+        if (!$mshipAccount) {
+            return Redirect::route('adm.mship.account.index');
+        }
+
+        // Let's try and load this RoleData
+        $role = RoleData::find(Input::get('role'));
+
+        if (!$role) {
+            return Redirect::route('adm.mship.account.details', [$mshipAccount->id, 'roles'])
+                ->withError('The selected role does not exist.');
+        }
+
+        // Let's add!
+        if (!$mshipAccount->roles->contains($role->id)) {
+            $mshipAccount->roles()
+                ->attach($role);
+        }
+
+        return Redirect::route('adm.mship.account.details', [$mshipAccount->id, 'roles'])
+            ->withSuccess($role->name . ' role attached successfully. This user inherited ' . count($role->permissions) . ' permissions.');
+    }
+}
